Move google recaptcha check to a concern
