+++ app/controllers/contact_requests_controller.rb
@@ -1,7 +1,7 @@
 class ContactRequestsController < ApplicationController
-  before_action :verify_recaptcha, only: :create
+  include Spammable
 
-  VALID_RECAPTCHA_SCORE = 0.5
+  before_action :verify_recaptcha, only: :create
 
   def new
     @contact_request = ContactRequest.new
@@ -27,26 +27,5 @@ class ContactRequestsController < ApplicationController
       :email,
       :message
     )
-  end
-
-  def verify_recaptcha
-    uri = URI("https://www.google.com/recaptcha/api/siteverify")
-    resp = Net::HTTP.post_form(
-      uri,
-      secret: ENV["GOOGLE_RECAPTCHA_SECRET_KEY"],
-      response: params["g-recaptcha-response"],
-      remoteip: request.remote_ip
-    )
-    parsed_json = JSON.parse(resp.body)
-    success = parsed_json["success"]
-    score = parsed_json["score"].to_f
-
-    unless success
-      Rails.error.report("Error with Google reCAPTCHA", context: parsed_json, handled: true)
-    end
-
-    if !success || score < VALID_RECAPTCHA_SCORE
-      redirect_to guest_root_path
-    end
   end
 end
